+++ src/StateSmith.Cli/Utils/TrackingConsole.cs
@@ -1,0 +1,20 @@
+using Spectre.Console;
+
+namespace StateSmith.Cli.Utils;
+
+public class TrackingConsole
+{
+    public bool Accessed = false;
+    private IAnsiConsole console;
+
+    public TrackingConsole(IAnsiConsole console)
+    {
+        this.console = console;
+    }
+
+    public IAnsiConsole Get()
+    {
+        Accessed = true;
+        return console;
+    }
+}